ai-agent-python/requirements.txt
fastapi
uvicorn
openai>=1.0.0
pydantic
python-dotenv
pytest
httpx
